Moved adding branch_key/app_key to central place in RemoteInterface.java
