+++ Branch-SDK/src/io/branch/referral/BranchRemoteInterface.java
@@ -1,6 +1,5 @@
 package io.branch.referral;
 
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -26,13 +25,12 @@ public class BranchRemoteInterface extends RemoteInterface {
 	public static final String REQ_TAG_SEND_APP_LIST = "t_send_app_list";
 
 	private SystemObserver sysObserver_;
-	private PrefHelper prefHelper_;
 	private NetworkCallback callback_;
 	
 	public BranchRemoteInterface() {}
 	
 	public BranchRemoteInterface(Context context) {
-		prefHelper_ = PrefHelper.getInstance(context);
+		super(context);
 		sysObserver_ = new SystemObserver(context);
 	}
 	
@@ -45,8 +43,6 @@ public class BranchRemoteInterface extends RemoteInterface {
 		if (callback_ != null) {
 			JSONObject installPost = new JSONObject();
 			try {
-				installPost.put(BRANCH_KEY, prefHelper_.getBranchKey());
-                installPost.put("app_id", prefHelper_.getAppKey());
 				if (!installID.equals(PrefHelper.NO_STRING_VALUE))
 					installPost.put("link_click_id", installID);
 				String uniqId = sysObserver_.getUniqueID(prefHelper_.getExternDebug());
@@ -102,8 +98,6 @@ public class BranchRemoteInterface extends RemoteInterface {
 		if (callback_ != null) {
 			JSONObject openPost = new JSONObject();
 			try {
-				openPost.put(BRANCH_KEY, prefHelper_.getBranchKey());
-                openPost.put("app_id", prefHelper_.getAppKey());
 				openPost.put("device_fingerprint_id", prefHelper_.getDeviceFingerPrintID());
 				openPost.put("identity_id", prefHelper_.getIdentityID());
 				openPost.put("is_referrable", prefHelper_.getIsReferrable());
@@ -135,8 +129,6 @@ public class BranchRemoteInterface extends RemoteInterface {
 		if (callback_ != null) {
 			JSONObject closePost = new JSONObject();
 			try {
-				closePost.put(BRANCH_KEY, prefHelper_.getBranchKey());
-                closePost.put("app_id", prefHelper_.getAppKey());
 				closePost.put("device_fingerprint_id", prefHelper_.getDeviceFingerPrintID());
 				closePost.put("identity_id", prefHelper_.getIdentityID());
 				closePost.put("session_id", prefHelper_.getSessionID());
@@ -172,28 +164,14 @@ public class BranchRemoteInterface extends RemoteInterface {
 	}
 	
 	public void getRewards() {
-		JSONObject post = new JSONObject();
-		try {
-			post.put(BRANCH_KEY, prefHelper_.getBranchKey());
-            post.put("app_id", prefHelper_.getAppKey());
-		} catch (JSONException ignore) {
-		}
-		String params = this.convertJSONtoString(post);
-		String urlExtend = "v1/credits/" + prefHelper_.getIdentityID() + params;
+		String urlExtend = "v1/credits/" + prefHelper_.getIdentityID();
 		if (callback_ != null) {
 			callback_.finished(make_restful_get(prefHelper_.getAPIBaseUrl() + urlExtend, REQ_TAG_GET_REWARDS, prefHelper_.getTimeout()));
 		}
 	}
 	
 	public void getReferralCounts() {
-		JSONObject post = new JSONObject();
-		try {
-			post.put(BRANCH_KEY, prefHelper_.getBranchKey());
-            post.put("app_id", prefHelper_.getAppKey());
-		} catch (JSONException ignore) {
-		}
-		String params = this.convertJSONtoString(post);
-		String urlExtend = "v1/referrals/" + prefHelper_.getIdentityID() + params;
+		String urlExtend = "v1/referrals/" + prefHelper_.getIdentityID();
 		if (callback_ != null) {
 			callback_.finished(make_restful_get(prefHelper_.getAPIBaseUrl() + urlExtend, REQ_TAG_GET_REFERRAL_COUNTS, prefHelper_.getTimeout()));
 		}
@@ -277,8 +255,6 @@ public class BranchRemoteInterface extends RemoteInterface {
 		try {
 			String urlExtend = "v1/debug/connect";
 			JSONObject post = new JSONObject();
-			post.put(BRANCH_KEY, prefHelper_.getBranchKey());
-            post.put("app_id", prefHelper_.getAppKey());
 			post.put("device_fingerprint_id", prefHelper_.getDeviceFingerPrintID());
 			if (sysObserver_.getBluetoothPresent()) {
 				post.put("device_name", BluetoothAdapter.getDefaultAdapter().getName());
@@ -299,8 +275,6 @@ public class BranchRemoteInterface extends RemoteInterface {
 		try {
 			String urlExtend = "v1/debug/disconnect";
 			JSONObject post = new JSONObject();
-			post.put(BRANCH_KEY, prefHelper_.getBranchKey());
-            post.put("app_id", prefHelper_.getAppKey());
 			post.put("device_fingerprint_id", prefHelper_.getDeviceFingerPrintID());
 			callback_.finished(make_restful_post(post, prefHelper_.getAPIBaseUrl() + urlExtend, PrefHelper.REQ_TAG_DEBUG_DISCONNECT, prefHelper_.getTimeout(), false));
 		} catch (JSONException ex) {
@@ -312,45 +286,11 @@ public class BranchRemoteInterface extends RemoteInterface {
 		try {
 			String urlExtend = "v1/debug/log";
 			JSONObject post = new JSONObject();
-			post.put(BRANCH_KEY, prefHelper_.getBranchKey());
-            post.put("app_id", prefHelper_.getAppKey());
 			post.put("device_fingerprint_id", prefHelper_.getDeviceFingerPrintID());
 			post.put("log", log);
 			callback_.finished(make_restful_post(post, prefHelper_.getAPIBaseUrl() + urlExtend, PrefHelper.REQ_TAG_DEBUG_LOG, prefHelper_.getTimeout(), false));
 		} catch (JSONException ex) {
 			ex.printStackTrace();
 		}
-	}
-	
-	private String convertJSONtoString(JSONObject json) {
-		StringBuilder result = new StringBuilder();
-		
-		if (json != null) {
-            JSONArray names = json.names();
-	        if (names != null) {
-                boolean first = true;
-                int size = names.length();
-                for(int i = 0; i < size; i++) {
-                	try {
-	                    String key = names.getString(i);
-	
-		        		if (first) {
-			        		result.append("?");
-			        		first = false;
-			        	} else {
-			        		result.append("&");
-			        	}
-	
-	                    String value = json.getString(key);
-	                    result.append(key).append("=").append(value);
-					} catch (JSONException e) {
-						e.printStackTrace();
-						return null;
-					}
-	        	}
-	        }
-	    }
-		
-		return result.toString();
 	}
 }
